Refactor: rename packages
